Enforce non-parallel building by setting MAKEFLAGS=-j1 globally
It seems that something in the build system is encouraging parallel
building more than in the past as I get rather often FTFBS with xymon
and the default settings (which should be non-parallel building with
debhelper compat level 9 and below).

Additionally, I now also get build failures in the install-server
target when using parallel builds, not only in the build target.